pyproject.toml
[tool.poetry]
name = "hackathon-social-media"
version = "0.1.0"
description = ""
authors = ["Nam Tran <nam3864779@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.9"
beautifulsoup4 = "^4.12.3"
langchain-core = "^0.3.15"
langchain-openai = "^0.2.6"
langgraph = "^0.2.45"
selenium = "^4.26.1"
rich = "^13.9.4"
langchain-anthropic = "^0.2.4"
whisper = "^1.1.10"
praw = "^7.8.1"
python-dotenv = "^1.0.1"
youtube-transcript-api = "^0.6.2"
gradio = "^5.5.0"
langserve = "^0.3.0"
pypdf2 = "^3.0.1"
playwright = "^1.48.0"
lxml = "^5.3.0"
sse-starlette = "^2.1.3"
tweepy = "^4.14.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
